moorefield: net: Fix setting tcp_default_init_rwnd
Theoretically this should improve network throughput.  The kernel
default is 10 but on boot init.rc sets this value (the initial
window size for congestion control) to 60.

Prior to this change it was impossible to change the value and it
would always remain set to 10 because for some reason Asus changed
the code to always restore the previous value (they changed:

   if (I_parsed_result_okay && value_is_not_in_valid_range)
      restore old value

to

   if (... || res <= 0 || ...)
      restore old value

where res == 0 on successful parsing and negative on error (aka
res <= 0 is always true).

Change-Id: I81ffcc5a59ac2a9c57fb60cb33d394b5e29f4495